+++ WalletOK/src/main/java/com/theost/walletok/presentation/wallet_details/transaction/widgets/TransactionValueListener.kt
@@ -2,6 +2,6 @@ package com.theost.walletok.presentation.wallet_details.transaction.widgets
 
 interface TransactionValueListener {
 
-    fun onValueSubmitted(value: Int)
+    fun onValueSubmitted(value: Long)
 
 }
